requirements.txt
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
botocore==1.37.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
db-dtypes==1.4.2
decorator==5.2.1
defusedxml==0.7.1
dlt==1.10.0
docopt==0.6.2
dotenv==0.9.9
executing==2.2.0
fastjsonschema==2.21.1
frozenlist==1.6.0
fsspec==2025.3.2
gcsfs==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
giturlparse==0.12.0
google-api-core==2.24.2
google-auth==2.40.0
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
hexbytes==1.3.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
makefun==1.15.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
multidict==6.4.3
narwhals==1.37.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.0.2
oauthlib==3.2.2
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathvalidate==3.2.3
pendulum==3.1.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
ply==3.11
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requirements-parser==0.11.0
rich==14.0.0
rich-argparse==1.7.0
rpds-py==0.24.0
rsa==4.9.1
s3fs==2025.3.2
semver==3.0.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sqlglot==26.16.4
stack-data==0.6.3
streamlit==1.45.0
tenacity==9.1.2
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
types-setuptools==80.1.0.20250503
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
yarg==0.1.9
yarl==1.20.0
zipp==3.21.0
